Moved copy paste code to Parent Class
Moved copy paste code from the Level Scenes to a new parent Class "GenericLevelScene" (less horrible)

Renamed Constants to GlobalValues since it now also contains the global variable for the current level (oh boi)
This is necessary so that the GenericLevelScene class is able to set the right background picture. Surely not the best solution, but I couldn't find any other way (yet).